refactoring, moving span tree construction to the SpanTree.scala
